* libmsip: W32 port:    - Removes warnings    - Renames SM_DEBUG (used in winows header files) to MSM_DEBUG
git-svn-id: svn://svn.minisip.org/minisip/trunk@1821 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmsip/source/headers/SipHeaderVia.cxx
@@ -50,7 +50,7 @@ SipHeaderValueVia::SipHeaderValueVia(const string &build_from)
 		: SipHeaderValue(SIP_HEADER_TYPE_VIA,sipHeaderValueViaTypeStr)
 {
 	unsigned i=0;
-	unsigned pos=0;
+	size_t pos=0;
 
 	ip="";
 	port=0;
